Server startup and routing complete for proxie server
Request handling logic needs to be filled in for v0 of the api.
Basic logging and all needed data is sent down to be handled by
the dcf functions. Currently there is no validation other than
checking the first two sections of the url paths against the
actions defined for the server QC handler in 'actions.js'